// src/types/json-record.types.ts
export type TJsonRecord = {
  datum: string;
  nummer1: string;
  nummer2: string;
  nummer3: string;
  nummer4: string;
  nummer5: string;
  zz1: string;
  zz2: string;
  spielEinsatz: string;
  anzahlKlasse1: string;
  quoteKlasse1: string;
  anzahlKlasse2: string;
  quoteKlasse2: string;
  anzahlKlasse3: string;
  quoteKlasse3: string;
  anzahlKlasse4: string;
  quoteKlasse4: string;
  anzahlKlasse5: string;
  quoteKlasse5: string;
  anzahlKlasse6: string;
  quoteKlasse6: string;
  anzahlKlasse7: string;
  quoteKlasse7: string;
  anzahlKlasse8: string;
  quoteKlasse8: string;
  anzahlKlasse9: string;
  quoteKlasse9: string;
  anzahlKlasse10: string;
  quoteKlasse10: string;
  anzahlKlasse11: string;
  quoteKlasse11: string;
  anzahlKlasse12: string;
  quoteKlasse12: string;
  tag: string;
};